+++ src/redturtle/volto/tests/test_advancedsearch.py
@@ -10,6 +10,7 @@ from redturtle.volto.interfaces import IRedTurtleVoltoSettings
 from redturtle.volto.testing import REDTURTLE_VOLTO_API_FUNCTIONAL_TESTING
 from transaction import commit
 
+import json
 import unittest
 
 
@@ -220,6 +221,44 @@ class AdvancedSearchTest(BaseTest):
         self.assertEqual(result["items_total"], 3)
         self.assertEqual(
             ["d1", "f1", "e1"], [item["@id"].split("/")[-1] for item in result["items"]]
+        )
+
+    def test_search_with_custom_rank(self):
+        query = {"SearchableText": "foo"}
+
+        # this is a query with default ranking values
+        result = self.api_session.get("/@search", params=query).json()
+        self.assertEqual(result["items_total"], 3)
+        self.assertEqual(
+            ["d1", "f1", "e1"], [item["@id"].split("/")[-1] for item in result["items"]]
+        )
+
+        # now we change ranking rules to have different order: add Event with
+        # weight 20 (higher than others)
+        api.portal.set_registry_record(
+            "advanced_query_ranking_rules",
+            json.dumps(
+                [
+                    {"index": "Subject", "value": "__TERM__", "weight": 16},
+                    {"index": "Title", "value": "__TERM__", "weight": 8},
+                    {"index": "Description", "value": "__TERM__", "weight": 6},
+                    {
+                        "index": "portal_type",
+                        "value": "Event",
+                        "weight": 20,
+                    },
+                ]
+            ),
+            interface=IRedTurtleVoltoSettings,
+        )
+
+        commit()
+
+        # re-execute query
+        result = self.api_session.get("/@search", params=query).json()
+        self.assertEqual(result["items_total"], 3)
+        self.assertEqual(
+            ["e1", "d1", "f1"], [item["@id"].split("/")[-1] for item in result["items"]]
         )